+++ R/utils.R
@@ -67,8 +67,8 @@ teal_data_to_filtered_data <- function(x, datanames = teal_data_datanames(x)) {
     x = sapply(datanames, function(dn) x[[dn]], simplify = FALSE),
     join_keys = teal.data::join_keys(x)
   )
-  # Piggy-back entire pre-processing code so that filtering code can be appended later.
-  attr(ans, "preprocessing_code") <- teal.code::get_code(x)
+  # Piggy-back pre-processing code for datasets of interest so that filtering code can be appended later.
+  attr(ans, "preprocessing_code") <- teal.data::get_code(x, datanames = datanames, check_names = FALSE)
   attr(ans, "verification_status") <- x@verified
   ans
 }
@@ -104,46 +104,6 @@ report_card_template <- function(title, label, description = NULL, with_filter, 
   card
 }
 
-#' Resolve `datanames` for the modules
-#'
-#' Modifies `module$datanames` to include names of the parent dataset (taken from `join_keys`).
-#' When `datanames` is set to `"all"` it is replaced with all available datasets names.
-#' @param modules (`teal_modules`) object
-#' @param datanames (`character`) names of datasets available in the `data` object
-#' @param join_keys (`join_keys`) object
-#' @return `teal_modules` with resolved `datanames`.
-#' @keywords internal
-resolve_modules_datanames <- function(modules, datanames, join_keys) {
-  if (inherits(modules, "teal_modules")) {
-    modules$children <- sapply(
-      modules$children,
-      resolve_modules_datanames,
-      simplify = FALSE,
-      datanames = datanames,
-      join_keys = join_keys
-    )
-    modules
-  } else {
-    modules$datanames <- if (identical(modules$datanames, "all")) {
-      datanames
-    } else if (is.character(modules$datanames)) {
-      extra_datanames <- setdiff(modules$datanames, datanames)
-      if (length(extra_datanames)) {
-        stop(
-          sprintf(
-            "Module %s has datanames that are not available in a 'data':\n %s not in %s",
-            modules$label,
-            toString(extra_datanames),
-            toString(datanames)
-          )
-        )
-      }
-      datanames_adjusted <- intersect(modules$datanames, datanames)
-      include_parent_datanames(dataname = datanames_adjusted, join_keys = join_keys)
-    }
-    modules
-  }
-}
 
 #' Check `datanames` in modules
 #'
@@ -217,6 +177,93 @@ check_filter_datanames <- function(filters, datanames) {
     paste(out, collapse = "\n")
   } else {
     TRUE
+  }
+}
+
+
+#' Create filterable data for modules
+#'
+#' Converts input data to a `FilteredData` object(s) to allow filtering before passing data to individual modules.
+#'
+#' @param data (`teal_data`)
+#' @param modules (`teal_modules`) object
+#' @param filters (`teal_slices`) object
+#' @param filtered_data_singleton A result of `teal_data_to_filtered_data` applied to `data`.
+#' @return Returns list of same shape as `modules`, containing `FilteredData` at every leaf.
+#' If module specific, each leaf contains different instance, otherwise every leaf contains `filtered_data_singleton`.
+#' @keywords internal
+modules_datasets <- function(data, modules, filters, filtered_data_singleton = teal_data_to_filtered_data(data)) {
+  checkmate::assert_class(data, "teal_data")
+  checkmate::assert_multi_class(modules, c("teal_modules", "teal_module"))
+  checkmate::assert_class(filters, "modules_teal_slices")
+  checkmate::assert_r6(filtered_data_singleton, "FilteredData")
+
+  if (!isTRUE(attr(filters, "module_specific"))) {
+    # subset global filters
+    slices <- shiny::isolate({
+      Filter(function(x) x$id %in% attr(filters, "mapping")$global_filters, filters)
+    })
+    filtered_data_singleton$set_filter_state(slices)
+    return(modules_structure(modules, filtered_data_singleton))
+  }
+
+  if (inherits(modules, "teal_module")) {
+    # 1. get datanames
+    datanames <-
+      if (is.null(modules$datanames) || identical(modules$datanames, "all")) {
+        include_parent_datanames(
+          teal_data_datanames(data),
+          teal.data::join_keys(data)
+        )
+      } else {
+        include_parent_datanames(
+          modules$datanames,
+          teal.data::join_keys(data)
+        )
+      }
+    # 2. subset filters (global + dedicated)
+    slices <- shiny::isolate({
+      Filter(x = filters, f = function(x) {
+        x$dataname %in% datanames &&
+          (x$id %in% attr(filters, "mapping")$global_filters ||
+            x$id %in% unique(unlist(attr(filters, "mapping")[modules$label]))) # nolint: indentation_linter.
+      })
+    })
+    # 2a. subset include/exclude varnames
+    slices$include_varnames <- attr(slices, "include_varnames")[names(attr(slices, "include_varnames")) %in% datanames]
+    slices$exclude_varnames <- attr(slices, "exclude_varnames")[names(attr(slices, "exclude_varnames")) %in% datanames]
+
+    # 3. instantiate FilteredData
+    filtered_data <- teal_data_to_filtered_data(data, datanames)
+    # 4. set state
+    filtered_data$set_filter_state(slices)
+    # 5. return
+    return(filtered_data)
+  } else if (inherits(modules, "teal_modules")) {
+    ans <- lapply(
+      modules$children,
+      modules_datasets,
+      data = data,
+      filters = filters,
+      filtered_data_singleton = filtered_data_singleton
+    )
+    names(ans) <- vapply(modules$children, `[[`, character(1), "label")
+
+    return(ans)
+  }
+
+  stop("something is not right")
+}
+
+# Returns nested list of same shape as `modules` with `value` at every leaf.
+modules_structure <- function(modules, value = TRUE) {
+  if (inherits(modules, "teal_module")) {
+    return(value)
+  } else {
+    stats::setNames(
+      lapply(modules$children, modules_structure, value),
+      vapply(modules$children, `[[`, character(1), "label")
+    )
   }
 }