+++ R/validate_inputs.R
@@ -41,7 +41,7 @@
 #'     sidebarPanel(
 #'       selectInput("letter", "select a letter:", c(letters[1:3], LETTERS[4:6])),
 #'       selectInput("number", "select a number:", 1:6),
-#'       br(),
+#'       tags$br(),
 #'       selectInput("color", "select a color:",
 #'         c("black", "indianred2", "springgreen2", "cornflowerblue"),
 #'         multiple = TRUE
@@ -58,12 +58,16 @@
 #'   # set up input validation
 #'   iv <- InputValidator$new()
 #'   iv$add_rule("letter", sv_in_set(LETTERS, "choose a capital letter"))
-#'   iv$add_rule("number", ~ if (as.integer(.) %% 2L == 1L) "choose an even number")
+#'   iv$add_rule("number", function(x) {
+#'     if (as.integer(x) %% 2L == 1L) "choose an even number"
+#'   })
 #'   iv$enable()
 #'   # more input validation
 #'   iv_par <- InputValidator$new()
 #'   iv_par$add_rule("color", sv_required(message = "choose a color"))
-#'   iv_par$add_rule("color", ~ if (length(.) > 1L) "choose only one color")
+#'   iv_par$add_rule("color", function(x) {
+#'     if (length(x) > 1L) "choose only one color"
+#'   })
 #'   iv_par$add_rule(
 #'     "size",
 #'     sv_between(
@@ -87,7 +91,7 @@
 #'       ))
 #'     )
 #'
-#'     plot(eruptions ~ waiting, faithful,
+#'     plot(faithful$eruptions ~ faithful$waiting,
 #'       las = 1, pch = 16,
 #'       col = input[["color"]], cex = input[["size"]]
 #'     )